requirements.txt
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
certifi
charset-normalizer
cloudscraper
idna
pyparsing
requests-toolbelt
